+++ core/security/application/doc.go
@@ -1,3 +1,1 @@
 package application
-
-//go:generate go run github.com/vektra/mockery/v2@v2.45.0 --name=SecurityService --case=underscore