+++ core/security/interface/middleware/doc.go
@@ -1,3 +1,1 @@
 package middleware
-
-//go:generate go run github.com/vektra/mockery/v2@v2.45.0 --all --case=underscore