+++ core/zap/module.go
@@ -12,7 +12,7 @@ import (
 )
 
 type (
-	// Module for logrus logging
+	// Module for zap logging
 	Module struct {
 		area               string
 		json               bool
@@ -24,11 +24,19 @@ type (
 		samplingThereafter float64
 		fieldMap           map[string]string
 		logSession         bool
+		callerEncoder      zapcore.CallerEncoder
+		callerskip         float64
 	}
 
 	shutdownEventSubscriber struct {
 		logger flamingo.Logger
 	}
+)
+
+const (
+	ZapCallerEncoderShort = "short"
+	ZapCallerEncoderSmart = "smart"
+	ZapCallerEncoderFull  = "full"
 )
 
 var logLevels = map[string]zapcore.Level{
@@ -40,6 +48,12 @@ var logLevels = map[string]zapcore.Level{
 	"DPanic": zap.DPanicLevel,
 	"Panic":  zap.PanicLevel,
 	"Fatal":  zap.FatalLevel,
+}
+
+var callerEncoders = map[string]zapcore.CallerEncoder{
+	ZapCallerEncoderSmart: smartCallerEncoder,
+	ZapCallerEncoderFull:  zapcore.FullCallerEncoder,
+	ZapCallerEncoderShort: zapcore.ShortCallerEncoder,
 }
 
 // Inject dependencies
@@ -54,6 +68,8 @@ func (m *Module) Inject(config *struct {
 	SamplingThereafter float64    `inject:"config:core.zap.sampling.thereafter,optional"`
 	FieldMap           config.Map `inject:"config:core.zap.fieldmap,optional"`
 	LogSession         bool       `inject:"config:core.zap.logsession,optional"`
+	CallerEncoder      string     `inject:"config:core.zap.encoding.caller,optional"`
+	CallerSkip         float64    `inject:"config:core.zap.callerskip,optional"`
 }) {
 	m.area = config.Area
 	m.json = config.JSON
@@ -64,6 +80,18 @@ func (m *Module) Inject(config *struct {
 	m.samplingInitial = config.SamplingInitial
 	m.samplingThereafter = config.SamplingThereafter
 	m.logSession = config.LogSession
+	m.callerEncoder = callerEncoders[ZapCallerEncoderShort]
+
+	// if not provided set 2 as default
+	m.callerskip = 2
+	if config.CallerSkip > 1 {
+		// otherwise let user config override
+		m.callerskip = config.CallerSkip
+	}
+
+	if encoder, ok := callerEncoders[config.CallerEncoder]; ok {
+		m.callerEncoder = encoder
+	}
 
 	if config.FieldMap != nil {
 		m.fieldMap = make(map[string]string, len(config.FieldMap))
@@ -76,7 +104,7 @@ func (m *Module) Inject(config *struct {
 	}
 }
 
-// Configure the zap logger as flamingo.Logger (in JSON mode kibana compatible)
+// Configure the zap logger as flamingo.Logger
 func (m *Module) Configure(injector *dingo.Injector) {
 	injector.Bind(new(flamingo.Logger)).ToInstance(m.createLoggerInstance())
 	flamingo.BindEventSubscriber(injector).To(shutdownEventSubscriber{})
@@ -128,7 +156,7 @@ func (m *Module) createLoggerInstance() *Logger {
 			EncodeLevel:    encoder,
 			EncodeTime:     zapcore.ISO8601TimeEncoder,
 			EncodeDuration: zapcore.SecondsDurationEncoder,
-			EncodeCaller:   zapcore.ShortCallerEncoder,
+			EncodeCaller:   m.callerEncoder,
 			EncodeName:     zapcore.FullNameEncoder,
 		},
 		OutputPaths:      []string{"stderr"},
@@ -136,7 +164,7 @@ func (m *Module) createLoggerInstance() *Logger {
 		InitialFields:    nil,
 	}
 
-	logger, err := cfg.Build(zap.AddCallerSkip(1))
+	logger, err := cfg.Build(zap.AddCallerSkip(int(m.callerskip)))
 	if err != nil {
 		panic(err)
 	}
@@ -186,8 +214,12 @@ core: zap: {
 	fieldmap: {
 		[string]: string
 	}
+	encoding: {
+		caller: *"%s" | "%s" | "%s"
+	}
+	callerskip: float64 | *2
 }
-`, allowedLevels)
+`, allowedLevels, ZapCallerEncoderShort, ZapCallerEncoderSmart, ZapCallerEncoderFull)
 }
 
 // FlamingoLegacyConfigAlias mapping