+++ CHANGELOG.md
@@ -1,5 +1,133 @@
 # Changelog
 
+## Version v3.16.0 (2025-08-06)
+
+### Features
+
+- **session:** add username support for redis (#524) (d503f108)
+
+### Chores and tidying
+
+- **deps:** bump github.com/go-viper/mapstructure/v2 (#517) (168a4e63)
+- **deps:** update module github.com/coreos/go-oidc/v3 to v3.15.0 (#523) (d68f6482)
+- **deps:** update module golang.org/x/sync to v0.16.0 (#519) (a6d3300e)
+
+## Version v3.15.0 (2025-06-13)
+
+### Features
+
+- zap logger caller skip configurable (#512) (6494dc09)
+
+### Chores and tidying
+
+- **deps:** update module github.com/redis/go-redis/v9 to v9.10.0 (#501) (56776d36)
+- **deps:** update dependency go to v1.24.4 (#496) (3dd1f7d5)
+- **deps:** update module github.com/vektra/mockery/v3 to v3.3.6 (#513) (d6f1605b)
+- **deps:** update module github.com/coreos/go-oidc/v3 to v3.14.1 (#497) (886be3a6)
+- **deps:** update module github.com/vektra/mockery/v2 to v3 (#508) (60afe694)
+- **deps:** update module golang.org/x/sync to v0.15.0 (#509) (dc9f9f88)
+- migrate golangci-lint to v2 (#506) (8e42afee)
+
+## Version v3.14.0 (2025-05-19)
+
+### Features
+
+- **healthcheck:** add measured status healthcheck interface (#503) (69049df2)
+- **web:** support Flush on http responses (#494) (e83a8c62)
+
+### Chores and tidying
+
+- **deps:** update module github.com/gorilla/sessions to v1.4.0 (#428) (3cc8e206)
+- **deps:** update module golang.org/x/oauth2 to v0.30.0 (#498) (b3368bea)
+- **deps:** update module golang.org/x/sync to v0.14.0 (#499) (e017aa5b)
+- **deps:** bump golang.org/x/net from 0.37.0 to 0.38.0 (#500) (57ea554d)
+- **deps:** update dependency go to v1.24.1 (#486) (712321fc)
+- **deps:** update module github.com/google/go-cmp to v0.7.0 (#482) (85d60c90)
+- **deps:** update module github.com/coreos/go-oidc/v3 to v3.13.0 (#489) (e5313988)
+- **deps:** update module github.com/redis/go-redis/v9 to v9.7.3 [security] (#491) (cbc03189)
+- **deps:** bump github.com/golang-jwt/jwt/v5 from 5.2.1 to 5.2.2 (#493) (eabcfb91)
+- **deps:** bump github.com/go-jose/go-jose/v4 from 4.0.2 to 4.0.5 (#484) (a7e5d1d4)
+- make post run script available for both Run and RunE functions (#478) (7bbfff73)
+
+### Other
+
+- Add tests for event dispatch when running commands and graceful shutdown (#485) (875204d2)
+
+## Version v3.13.0 (2025-02-17)
+
+### Features
+
+- **core/zap:** Make caller encoder configurable, add additional one (#301) (fc037fce)
+
+### Chores and tidying
+
+- **deps:** update module github.com/spf13/cobra to v1.9.1 (#477) (5fad27fd)
+- **deps:** update module golang.org/x/sync to v0.11.0 (#474) (5800854f)
+- **deps:** update module golang.org/x/oauth2 to v0.26.0 (#473) (48e6b955)
+- **deps:** update module github.com/spf13/pflag to v1.0.6 (#471) (5667c5d3)
+- **deps:** bump golang.org/x/net from 0.27.0 to 0.33.0 (#467) (45c1368c)
+- bump go version with mockery and linter (#476) (98b92505)
+- edit description of sessions module (#475) (fe338e85)
+- **deps:** update module github.com/vektra/mockery/v2 to v2.51.1 (#468) (caa5d915)
+
+## Version v3.12.0 (2025-01-09)
+
+### Features
+
+- **app:** add dingo-trace-injections flag, which switches injection tracing (#455) (c25a60ee)
+- **web/handler:** add http method to opencensus tag for flamingo_router_controller (#432) (7597e699)
+
+### Chores and tidying
+
+- **deps:** update module github.com/coreos/go-oidc/v3 to v3.12.0 (#465) (e40d9eda)
+- **deps:** update module github.com/vektra/mockery/v2 to v2.50.4 (#464) (7d2d7f05)
+- **deps:** update module golang.org/x/oauth2 to v0.25.0 (#466) (27262c8a)
+- **docs:** add troubleshooting information (#457) (8426b77e)
+- edit mockery configuration, remove scattered go:generate annotations, edited linter config (#463) (586b32ba)
+- **deps:** update dependency go to v1.23.4 (#459) (a307b167)
+- **deps:** bump golang.org/x/crypto from 0.25.0 to 0.31.0 (#462) (ad862f2d)
+- **deps:** update module golang.org/x/sync to v0.10.0 (#461) (e1b39ea6)
+- **deps:** update module golang.org/x/oauth2 to v0.24.0 (#452) (a90e42b2)
+- **deps:** update module golang.org/x/sync to v0.9.0 (#453) (acde3c2b)
+- **deps:** update module github.com/vektra/mockery/v2 to v2.49.1 (#454) (af56c07a)
+- **deps:** update dependency go to v1.23.3 (#451) (77cf423b)
+- **deps:** update module github.com/redis/go-redis/v9 to v9.7.0 (#450) (0569eadc)
+- **deps:** update dependency go to v1.23.2 (#447) (7d64c4e4)
+- **deps:** update module github.com/vektra/mockery/v2 to v2.46.3 (#448) (2871fb36)
+- **deps:** update quay.io/keycloak/keycloak docker tag to v25.0.6 (#443) (be287200)
+- **deps:** update module go.uber.org/automaxprocs to v1.6.0 (#444) (9de9569f)
+- **deps:** update module github.com/vektra/mockery/v2 to v2.46.1 (#445) (5158f5a6)
+
+### Other
+
+- Add version command to the default framework CLI commands (#458) (7903a65b)
+
+## Version v3.11.0 (2024-09-19)
+
+### Features
+
+- decouple opencensus from core (#442) (95e83305)
+
+### Fixes
+
+- **cache:** pass deadline to context in load function (#440) (41c5fe24)
+
+### Chores and tidying
+
+- **deps:** update quay.io/keycloak/keycloak docker tag to v25.0.5 (#439) (634092be)
+- **deps:** update dependency go to v1.23.1 (#437) (e9e669e0)
+- **deps:** update module github.com/vektra/mockery/v2 to v2.46.0 (#438) (39733f67)
+
+## Version v3.10.1 (2024-09-05)
+
+### Fixes
+
+- fix file session store creation, if directory exists (#433) (6469f207)
+
+### Chores and tidying
+
+- **deps:** update module golang.org/x/oauth2 to v0.23.0 (#435) (c4a1fe51)
+
 ## Version v3.10.0 (2024-09-02)
 
 ### Features